JAMES-2214 Rename ambiguous and misleading MailboxRightsException
MailboxRightsException leads to think it concerns mailbox rights, but was raised when having rights on a delegated mailbox
